requirements.txt
eli5==0.13.0
fuzzywuzzy==0.18.0
gensim==4.3.2
ipykernel==6.25.2
Levenshtein==0.23.0
lime==0.2.0.1
matplotlib==3.8.0
mlflow==2.8.0
nltk==3.8.1
numpy==1.26.1
pandas==2.1.1
PyALE==1.1.3
pygam==0.9.0
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.13.0
sklearn-features==0.0.2
spacy==3.7.2
shap==0.43.0
xgboost==2.0.1
XPER==0.0.72

# linting
pre-commit==3.3.3
black==22.10.0
black[jupyter]
isort==5.12.0
nbstripout==0.6.1
ruff==0.0.272
